+++ src/compas/data/__init__.py
@@ -8,7 +8,6 @@ from .validators import is_int3
 from .validators import is_float3
 from .validators import is_float4x4
 from .validators import is_item_iterable
-from .validators import validate_data
 from .encoders import DataEncoder
 from .encoders import DataDecoder
 from .data import Data
@@ -31,5 +30,4 @@ __all__ = [
     "json_loads",
     "json_dump",
     "json_dumps",
-    "validate_data",
 ]